Extract out local cluster test from bench-tps
