+++ bench-tps/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod bench;
+pub mod cli;
