+++ local_cluster/Cargo.toml
@@ -11,9 +11,13 @@ homepage = "https://solana.com/"
 [dependencies]
 log = "0.4.8"
 rand = "0.6.5"
+solana-bench-tps = { path = "../bench-tps", version = "0.19.0-pre0" }
 solana-core = { path = "../core", version = "0.19.0-pre0" }
 solana-client = { path = "../client", version = "0.19.0-pre0" }
+solana-drone = { path = "../drone", version = "0.19.0-pre0" }
 solana-logger = { path = "../logger", version = "0.19.0-pre0" }
+solana-move-loader-api = { path = "../programs/move_loader_api", version = "0.19.0-pre0" }
+solana-move-loader-program = { path = "../programs/move_loader_program", version = "0.19.0-pre0" }
 solana-runtime = { path = "../runtime", version = "0.19.0-pre0" }
 solana-sdk = { path = "../sdk", version = "0.19.0-pre0" }
 solana-stake-api = { path = "../programs/stake_api", version = "0.19.0-pre0" }
